+++ alpha/external-web/app/routes.js
@@ -1,15 +1,13 @@
 var express = require('express')
 var mongo = require('./database')
 var router = express.Router()
-var multer = require('multer')
-
-var upload = multer({ dest: 'eligibility-uploads/' })
 
 module.exports = router
 
 // Include custom route files.
 require('./relationship');
 require('./escorts');
+require('./about-your-income');
 require('./application-submitted');
 
 /**
@@ -17,13 +15,6 @@ require('./application-submitted');
  */
 router.get('/', function (request, response) {
   response.render('index')
-})
-
-/**
- * Render the about you income page.
- */
-router.get('/about-your-income', function (request, response) {
-  response.render('about-your-income')
 })
 
 /**
@@ -35,29 +26,4 @@ router.post('/about-you', function (request, response) {
       response.render('add-user-success')
     }
   })
-})
-
-/**
- * Upload a single file to the system and it's meta data to the database.
- */
-router.post('/about-your-income', upload.single('evidence'), function (request, response) {
-  if (request.file) {
-    var metadata = {
-      eligibilityId: request.file.filename, // using filename guild for temp id
-      claimantID: null, // TODO: This should tie to a real claimant id.
-      originalFilename: request.file.originalname,
-      path: request.file.path
-    }
-
-    // Save the uploaded files meta data to the database.
-    mongo.db.collection('supporting-document').insertOne(metadata, function (error, result) {
-      if (!error) {
-        console.log('Successfully saved file meta data.')
-      }
-    })
-  }
-
-  // TODO: validation
-
-  response.redirect('/application-submitted')
 })
